requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.3
sentence-transformers==2.5.1
scikit-learn==1.4.0
plotly==5.18.0
geopy==2.4.1
folium==0.15.1
streamlit-folium==0.15.1
nltk==3.8.1
pycountry==23.12.11
wikipedia==1.4.0
requests==2.31.0
python-dotenv==1.0.1 